requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
classes==0.1.0
click==8.0.1
discord==1.7.3
discord.py==1.7.3
Flask==2.0.1
flatbuffers==1.12
fpdf==1.7.2
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
idna==3.2
img2pdf==0.4.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
multidict==5.1.0
multitasking==0.0.9
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.1
pikepdf==2.16.1
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
selenium==3.141.0
simplejson==3.17.4
six==1.15.0
sklearn==0.0
soup2dict==1.0.1
soupsieve==2.2.1
termcolor==1.1.0
threadpoolctl==2.2.0
tqdm==4.62.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3
